requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
PyMuPDF
cohere
pinecone-client
python-dotenv
requests
sqlalchemy
psycopg2